requirements.txt
gunicorn~=23.0
Flask~=3.1
Flask-Minify~=0.49
flask-babel~=4.0
pyotp~=2.9
pytz~=2025.1
python-dotenv~=1.0
qrcode~=8.0